fixing swinger shape

+++ library/swinger.c
@@ -14,10 +14,10 @@ const double INCREMENT = 0.1;
 const double SWING_EPSILON = 0.01;
 const int SWINGER_MASS = 3;
 const int GRAVITY = 10;
-const int SEMICIRCLE_MIN = M_PI/2;
-const int SEMICIRCLE_MAX = 3*M_PI/2;
-const int LEFT_SWINGER_ANGLE = 11*M_PI/6;
-const int RIGHT_SWINGER_ANGLE = 7*M_PI/6;
+const double SEMICIRCLE_MIN = M_PI/2;
+const double SEMICIRCLE_MAX = 3*M_PI/2;
+const double LEFT_SWINGER_ANGLE = 11*M_PI/6;
+const double RIGHT_SWINGER_ANGLE = 7*M_PI/6;
 const int MOMENTUM_EXPONENT = 3;
 const int QUARTER_ANGLE = M_PI/2;
 
@@ -30,7 +30,7 @@ typedef struct swinger {
     double length; // distance from swinger center to outer point
     double torque;
     vector_t force;
-    double momentum; // of the angular variety :)
+    double momentum; // angular
 } swinger_t;
 
 list_t *make_shape(vector_t center, double angle, double length){
